requirements.txt
aws-cdk-lib==2.12.0
constructs>=10.0.0,<11.0.0
cdk-lambda-powertools-python-layer